fix db get fork connection + redirect wps-xml/kvp to wps-rest job
