+++ weaver/wps/views.py
@@ -2,198 +2,11 @@
 pywps 4.x wrapper
 """
 import logging
-import os
-from typing import TYPE_CHECKING
-
-from configparser import ConfigParser
-from owslib.wps import WPSExecution
-from pyramid.httpexceptions import HTTPBadRequest, HTTPSeeOther
-from pyramid.settings import asbool
-from pyramid.threadlocal import get_current_request
 from pyramid.wsgi import wsgiapp2
-from pyramid_celery import celery_app as app
-from pywps.app import WPSRequest
-from pywps.app.Service import Service
-from pywps.response import WPSResponse
-from pywps.response.execute import ExecuteResponse
 
-from weaver.database import get_db
-from weaver.formats import CONTENT_TYPE_ANY_XML, CONTENT_TYPE_APP_JSON
-from weaver.owsexceptions import OWSNoApplicableCode
-from weaver.processes.convert import wps2json_io
-from weaver.processes.execution import submit_job_handler
-from weaver.processes.types import PROCESS_WORKFLOW
-from weaver.processes.utils import get_process
-from weaver.store.base import StoreProcesses
-from weaver.utils import get_header, get_settings, get_weaver_url
-from weaver.wps.utils import check_wps_status, get_wps_path, load_pywps_config
-from weaver.wps_restapi.processes.processes import get_job_submission_response
-from weaver.wps_restapi.swagger_definitions import BadRequestGetProcessInfoResponse, process_uri, processes_uri
-from weaver.visibility import VISIBILITY_PUBLIC
+from weaver.wps.service import get_pywps_service
 
 LOGGER = logging.getLogger(__name__)
-if TYPE_CHECKING:
-    from weaver.typedefs import AnySettingsContainer        # noqa: F401
-    from typing import AnyStr, Dict, Union, Optional        # noqa: F401
-
-
-class WorkerService(Service):
-    """
-    Dispatches PyWPS requests from *older* WPS-1/2 XML endpoint to WPS-REST as appropriate.
-
-    .. note::
-        For every WPS-Request type, the parsing of XML content is already handled by the PyWPS service for GET/POST.
-
-    When ``GetCapabilities`` or ``DescribeProcess`` requests are received, directly return to result as XML based
-    on content (no need to subprocess as Celery task that gets resolved quickly with only the process(es) details).
-    When JSON content is requested, instead return the redirect link to corresponding WPS-REST API endpoint.
-
-    When receiving ``Execute`` request, convert the XML payload to corresponding JSON and
-    dispatch it to the Celery Worker to actually process it after job setup for monitoring.
-    """
-    def __init__(self, *_, is_worker=False, settings=None, **__):
-        super(WorkerService, self).__init__(*_, **__)
-        self.is_worker = is_worker
-        self.settings = settings or get_settings(app)
-
-    def get_capabilities(self, wps_request, *_, **__):
-        """
-        Redirect to WPS-REST endpoint if requested ``Content-Type`` is JSON, or handle ``GetCapabilities`` normally.
-        """
-        req = wps_request.http_request
-        accept_type = get_header("Accept", req.headers)
-        if accept_type == CONTENT_TYPE_APP_JSON:
-            url = get_weaver_url(self.settings)
-            resp = HTTPSeeOther(location="{}{}".format(url, processes_uri))  # redirect
-            setattr(resp, "_update_status", lambda *_, **__: None)  # patch to avoid pywps server raising
-            return resp
-        return super(WorkerService, self).get_capabilities(wps_request, *_, **__)
-
-    def describe(self, wps_request, *_, **__):
-        """
-        Redirect to WPS-REST endpoint if requested ``Content-Type`` is JSON, or handle ``DescribeProcess`` normally.
-        """
-        req = wps_request.http_request
-        accept_type = get_header("Accept", req.headers)
-        if accept_type == CONTENT_TYPE_APP_JSON:
-            url = get_weaver_url(self.settings)
-            proc = wps_request.identifiers
-            if not proc:
-                raise HTTPBadRequest(BadRequestGetProcessInfoResponse.description)
-            if len(proc) > 1:
-                raise HTTPBadRequest("Unsupported multi-process ID for description. Only provide one.")
-            path = process_uri.format(process_id=proc[0])
-            resp = HTTPSeeOther(location="{}{}".format(url, path))  # redirect
-            setattr(resp, "_update_status", lambda *_, **__: None)  # patch to avoid pywps server raising
-            return resp
-        return super(WorkerService, self).describe(wps_request, *_, **__)
-
-    def execute(self, identifier, wps_request, uuid):
-        # type: (AnyStr, WPSRequest, AnyStr) -> WPSResponse
-        """
-        Dispatch operation to WPS-REST endpoint, which in turn should call back the real Celery Worker for execution.
-
-        Overrides the original execute operation, that instead will get handled by :meth:`execute_job` following
-        callback from Celery Worker that will handle process job creation and monitoring.
-        """
-        req = wps_request.http_request
-        pid = req.args.get("identifier")
-        proc = get_process(process_id=pid, settings=self.settings)  # raises if invalid or missing
-
-        # create the JSON payload from the XML content and submit job
-        is_workflow = proc.type == PROCESS_WORKFLOW
-        tags = req.args.get("tags", "").split(",") + ["xml", "wps-{}".format(wps_request.version)]
-        data = {
-            "inputs": [wps2json_io(i) for i in wps_request.inputs],
-            "outputs": [wps2json_io(o) for o in wps_request.outputs],
-            "response": EXECUTE_RESPONSE_DOCUMENT,
-            "mode": EXECUTE_MODE_ASYNC,
-        }
-        for output in data["outputs"]:
-            output["transmissionMode"] = EXECUTE_TRANSMISSION_MODE_REFERENCE
-        body = submit_job_handler(data, self.settings, proc.processEndpointWPS1,
-                                  process_id=pid, is_local=True, is_workflow=is_workflow, visibility=VISIBILITY_PUBLIC,
-                                  language=wps_request.language, tags=tags, auth=dict(req.headers))
-
-        # if accept was JSON, provide response content as is
-        accept_type = get_header("Accept", req.headers)
-        if accept_type == CONTENT_TYPE_APP_JSON:
-            setattr(resp, "_update_status", lambda *_, **__: None)  # patch to avoid pywps server raising
-            return get_job_submission_response(body)
-
-        # otherwise, recreate the equivalent content with expected XML template format
-        data = resp.json_body
-        job_url = data["location"]
-        wps_url = job_url.split("/jobs")[0]
-        job_id = data["jobID"]
-
-        class WorkerExecuteResponse(ExecuteResponse):
-            def __init__(self, *_, **__):
-                self.wps_request = wps_request
-                self.wps_request.raw = None  # make sure the doc gets generated
-                self.uuid = job_id
-                self.message = "Process {} accepted".format(pid)
-
-            @property
-            def json(self):
-                data.update({
-                    "language": self.wps_request.language,
-                    "service_instance": wps_url,
-                    "status": self._process_accepted(),
-                    "status_location": job_url,
-                })
-                return data
-
-        # when called by the WSGI app, '__call__' will generate the XML from 'doc',
-        # which itself is generated by template substitution from the data obtained by 'json'
-        return WorkerExecuteResponse()
-
-    def execute_job(self, process_id, wps_inputs, wps_outputs, mode, job_uuid):
-        """
-        Real execution of the process by active Celery Worker.
-        """
-        execution = WPSExecution(version="2.0", url="localhost")
-        xml_request = execution.buildRequest(process_id, wps_inputs, wps_outputs, mode=mode, lineage=True)
-        wps_request = WPSRequest()
-        wps_request.identifier = process_id
-        wps_request.set_version("2.0.0")
-        request_parser = wps_request._post_request_parser(wps_request.WPS.Execute().tag)
-        request_parser(xml_request)
-        # NOTE:
-        #  Setting 'status = false' will disable async execution of 'pywps.app.Process.Process'
-        #  but this is needed since this job is running within Celery already async
-        #  (daemon process can't have children processes)
-        #  Because if how the code in PyWPS is made, we have to re-enable creation of status file
-        wps_request.status = "false"
-        wps_response = super(WorkerService, self).execute(process_id, wps_request, job_uuid)
-        wps_response.store_status_file = True
-        # update execution status with actual status file and apply required references
-        execution = check_wps_status(location=wps_response.process.status_location, settings=self.settings)
-        execution.request = xml_request
-        return execution
-
-
-def get_pywps_service(environ=None, is_worker=False):
-    """
-    Generates the PyWPS Service that provides *older* WPS-1/2 XML endpoint.
-    """
-    environ = environ or {}
-    try:
-        # get config file
-        settings = get_settings(app)
-        pywps_cfg = environ.get("PYWPS_CFG") or settings.get("PYWPS_CFG") or os.getenv("PYWPS_CFG")
-        if not isinstance(pywps_cfg, ConfigParser) or not settings.get("weaver.wps_configured"):
-            load_pywps_config(app, config=pywps_cfg)
-
-        # call pywps application with processes filtered according to the adapter's definition
-        process_store = get_db(app).get_store(StoreProcesses)
-        processes_wps = [process.wps() for process in
-                         process_store.list_processes(visibility=VISIBILITY_PUBLIC, request=get_current_request())]
-        service = WorkerService(processes_wps, is_worker=is_worker, settings=settings)
-    except Exception as ex:
-        LOGGER.exception("Error occurred during PyWPS Service and/or Processes setup.")
-        raise OWSNoApplicableCode("Failed setup of PyWPS Service and/or Processes. Error [{!r}]".format(ex))
-    return service
 
 
 # @app.task(bind=True)
